# swagger-spec.json
{"openapi":"3.0.0","paths":{"/users/{id}":{"get":{"operationId":"UsersController_get","summary":"Get a user","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"The user has been successfully retrieved."},"403":{"description":"Forbidden."}},"tags":["users"]},"patch":{"operationId":"UsersController_update","summary":"Update a user","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserDto"}}}},"responses":{"200":{"description":"The user has been successfully updated."},"403":{"description":"Forbidden."}},"tags":["users"]},"delete":{"operationId":"UsersController_delete","summary":"Delete a user","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"The user has been successfully deleted."},"403":{"description":"Forbidden."}},"tags":["users"]}},"/users":{"get":{"operationId":"UsersController_list","summary":"List all users","parameters":[{"name":"limit","required":false,"in":"query","schema":{"minimum":1,"type":"number"}},{"name":"offset","required":false,"in":"query","schema":{"minimum":1,"type":"number"}}],"responses":{"200":{"description":"All users have been successfully retrieved."},"403":{"description":"Forbidden."}},"tags":["users"]},"post":{"operationId":"UsersController_create","summary":"Create a user","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"201":{"description":"The user has been successfully created."},"403":{"description":"Forbidden."}},"tags":["users"]}},"/companies/{id}":{"get":{"operationId":"CompaniesController_get","summary":"Get a company","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"The company has been successfully retrieved."},"403":{"description":"Forbidden."}},"tags":["companies"]},"patch":{"operationId":"CompaniesController_update","summary":"Update a company","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateCompanyDto"}}}},"responses":{"200":{"description":"The company has been successfully updated."},"403":{"description":"Forbidden."}},"tags":["companies"]},"delete":{"operationId":"CompaniesController_delete","summary":"Delete a company","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"The company has been successfully deleted."},"403":{"description":"Forbidden."}},"tags":["companies"]}},"/companies":{"get":{"operationId":"CompaniesController_list","summary":"List all companies","parameters":[{"name":"limit","required":false,"in":"query","schema":{"minimum":1,"type":"number"}},{"name":"offset","required":false,"in":"query","schema":{"minimum":1,"type":"number"}}],"responses":{"200":{"description":"The companies have been successfully retrieved."},"403":{"description":"Forbidden."}},"tags":["companies"]},"post":{"operationId":"CompaniesController_create","summary":"Create a company","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateCompanyDto"}}}},"responses":{"201":{"description":"The company has been successfully created."},"403":{"description":"Forbidden."}},"tags":["companies"]}},"/authentication/sign-up":{"post":{"operationId":"AuthenticationController_signUp","summary":"Sign a user up","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignUpDto"}}}},"responses":{"201":{"description":"User has been successfully signed up."}},"tags":["authentication"]}},"/authentication/sign-in":{"post":{"operationId":"AuthenticationController_signIn","summary":"Sign a user in","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignInDto"}}}},"responses":{"200":{"description":"User has been successfully signed in."},"401":{"description":"Unauthorized."}},"tags":["authentication"]}},"/authentication/refresh-tokens":{"post":{"operationId":"AuthenticationController_refreshTokens","summary":"Refresh user tokens","parameters":[],"responses":{"200":{"description":"Access token and refresh token have been successfully renewed."},"401":{"description":"Unauthorized."}},"tags":["authentication"]}}},"info":{"title":"Biblio API","description":"Biblio API documentation","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"schemas":{"CreateUserDto":{"type":"object","properties":{"firstname":{"type":"string","description":"The firstname of a user."},"lastname":{"type":"string","description":"The lastname of a user."},"email":{"type":"string","description":"The email of a user."},"password":{"type":"string","minLength":12,"description":"The password of a user."},"companyId":{"type":"string","description":"The company ID of a user."}},"required":["firstname","lastname","email","password"]},"UpdateUserDto":{"type":"object","properties":{"firstname":{"type":"string","description":"The firstname of a user."},"lastname":{"type":"string","description":"The lastname of a user."},"email":{"type":"string","description":"The email of a user."},"password":{"type":"string","minLength":12,"description":"The password of a user."},"companyId":{"type":"string","description":"The company ID of a user."}}},"CreateCompanyDto":{"type":"object","properties":{"name":{"type":"string","description":"The name of a company."},"address":{"type":"string","description":"The address of a company."},"postcode":{"type":"string","description":"The post code of a company."},"city":{"type":"string","description":"The city of a company."},"country":{"type":"string","description":"The country of a company."}},"required":["name","address","postcode","city","country"]},"UpdateCompanyDto":{"type":"object","properties":{"name":{"type":"string","description":"The name of a company."},"address":{"type":"string","description":"The address of a company."},"postcode":{"type":"string","description":"The post code of a company."},"city":{"type":"string","description":"The city of a company."},"country":{"type":"string","description":"The country of a company."}}},"SignUpDto":{"type":"object","properties":{"firstname":{"type":"string","description":"The firstname of a user."},"lastname":{"type":"string","description":"The lastname of a user."},"email":{"type":"string","description":"The email of a user."},"password":{"type":"string","minLength":12,"description":"The password of a user."}},"required":["firstname","lastname","email","password"]},"SignInDto":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string","minLength":12}},"required":["email","password"]}}}}